fix: address Copilot review feedback on template scaffolding
- Extract shared IPC types (CreateFailureReason, CreateTemplateResult,
  TemplateLogEntry) into src/shared/templates.ts so main, preload, and
  renderer import from a single source of truth instead of duplicating
  union literals.

- Use full validateProjectName() instead of just PROJECT_NAME_REGEX in
  the main-process createNextAppTemplate guard, so reserved names like
  "node_modules" and the length cap are enforced server-side too.

- Fix comment on PROJECT_NAME_REGEX to clarify it is a necessary but
  not sufficient subset check, not the authoritative validator.

- Add explicit empty-name guard in handleCreate() so an empty input
  cannot slip past the nameIssue === null check on submit.

- Distinguish "already added as project" vs "folder exists on disk"
  with separate PathStatus values and distinct i18n strings.

- Add @shared alias to the renderer Vitest project so renderer tests
  that import from @shared/ resolve correctly.
